requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
faiss-cpu
pandas 
streamlit
numpy
pytest
deepeval